requirements.txt
psycopg2-binary
PyQt6
matplotlib
docker
